+++ keewe-api/src/main/java/ccc/keeweapi/service/nest/PostApiService.java
@@ -2,7 +2,7 @@ package ccc.keeweapi.service.nest;
 
 import ccc.keeweapi.dto.nest.*;
 import ccc.keewedomain.domain.nest.AnnouncementPost;
-import ccc.keewedomain.service.PostDomainService;
+import ccc.keewedomain.service.nest.PostDomainService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -23,7 +23,7 @@ public class PostApiService {
 
     @Transactional
     public AnnouncementCreateResponse createAnnouncementPost(AnnouncementCreateRequest request) {
-        AnnouncementPost post = postDomainService.save(postAssembler.toAnnouncementCreateDto(request));
+        AnnouncementPost post = postDomainService.createAnnouncementPost(postAssembler.toAnnouncementCreateDto(request));
         return postAssembler.toAnnouncementCreateResponse(post);
     }
 }
