+++ keewe-api/src/main/java/ccc/keeweapi/service/user/ProfileService.java
@@ -7,8 +7,8 @@ import ccc.keewedomain.domain.common.enums.Activity;
 import ccc.keewedomain.domain.user.Profile;
 import ccc.keewedomain.domain.user.SocialLink;
 import ccc.keewedomain.repository.user.ProfileRepository;
-import ccc.keewedomain.service.ProfileDomainService;
-import ccc.keewedomain.service.SocialLinkDomainService;
+import ccc.keewedomain.service.user.ProfileDomainService;
+import ccc.keewedomain.service.user.SocialLinkDomainService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
