+++ keewe-api/src/main/java/ccc/keeweapi/service/user/UserApiService.java
@@ -4,8 +4,8 @@ import ccc.keeweapi.config.security.jwt.JwtUtils;
 import ccc.keeweapi.dto.user.UserAssembler;
 import ccc.keeweapi.dto.user.UserSignUpResponse;
 import ccc.keewedomain.domain.user.User;
-import ccc.keewedomain.service.ProfileDomainService;
-import ccc.keewedomain.service.UserDomainService;
+import ccc.keewedomain.service.user.ProfileDomainService;
+import ccc.keewedomain.service.user.UserDomainService;
 import ccc.keeweinfra.vo.OauthAccount;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
