+++ keewe-api/src/test/java/ccc/keeweapi/security/WithKeeweUserSecurityFactory.java
@@ -2,7 +2,7 @@ package ccc.keeweapi.security;
 
 import ccc.keeweapi.config.security.UserPrincipal;
 import ccc.keewedomain.domain.user.User;
-import ccc.keewedomain.dto.UserSignUpDto;
+import ccc.keewedomain.dto.user.UserSignUpDto;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
