+++ keewe-domain/src/main/java/ccc/keewedomain/service/nest/PostDomainService.java
@@ -1,13 +1,14 @@
-package ccc.keewedomain.service;
+package ccc.keewedomain.service.nest;
 
 import ccc.keewedomain.domain.nest.AnnouncementPost;
 import ccc.keewedomain.domain.nest.Candidate;
 import ccc.keewedomain.domain.nest.Post;
 import ccc.keewedomain.domain.nest.VotePost;
 import ccc.keewedomain.domain.user.Profile;
-import ccc.keewedomain.dto.VotePostDto;
+import ccc.keewedomain.dto.nest.VotePostDto;
 import ccc.keewedomain.dto.nest.AnnouncementCreateDto;
 import ccc.keewedomain.repository.nest.PostRepository;
+import ccc.keewedomain.service.user.ProfileDomainService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -37,14 +38,13 @@ public class PostDomainService {
         return ((T) postRepository.save(post)).getId();
     }
 
-
     private List<Candidate> createVoteCandidates(List<String> candidates, VotePost votePost) {
         return candidates.stream()
                 .map(it -> Candidate.from(it, votePost))
                 .collect(Collectors.toList());
     }
 
-    public AnnouncementPost save(AnnouncementCreateDto dto) {
+    public AnnouncementPost createAnnouncementPost(AnnouncementCreateDto dto) {
         Profile profile = profileDomainService.getAndVerifyOwnerOrElseThrow(dto.getProfileId(), dto.getUserId());
         return (AnnouncementPost) postRepository.save(AnnouncementPost.of(profile, dto.getContent()));
     }
